requirements.txt
--extra-index-url https://code.oak-tree.tech/api/v4/projects/335/packages/pypi/simple

fastapi==0.76.0
celery==5.2.6
alembic==1.7.7
httpx==0.15.5
psycopg2==2.9.3
pytest==6.1.0
requests==2.24.0
SQLAlchemy==1.4.36
uvicorn[standard]==0.17.6
bcrypt==3.2.0
sqlalchemy-utils==0.38.2
python-multipart==0.0.5
pyjwt==1.7.1
pandas==1.4.2
recordlinkage==0.15
phonetics==1.0.5
nltk==3.7
kafka-python==2.0.2
fusionauth-client==1.36.0
Jinja2==3.1.2
fastapi-utils==0.2.1
apscheduler==3.9.1
sse_starlette

client==0.2.2
sonador==0.2.2a2